backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-groq
langchain-community
sqlalchemy
pydantic
python-dotenv
asyncpg
httpx
